+++ pollbot/telegram/message_handler.py
@@ -3,8 +3,9 @@ from pollbot.display import get_settings_text
 from pollbot.enums import ExpectedInput, PollType, ReferenceType
 from pollbot.i18n import i18n
 from pollbot.models import Reference
-from pollbot.poll.helper import remove_old_references, init_votes_for_new_options
-from pollbot.poll.option import add_options_multiline, next_option
+from pollbot.poll.helper import remove_old_references
+from pollbot.poll.option import add_options_multiline
+from pollbot.display.poll.option import next_option
 from pollbot.poll.update import update_poll_messages
 from pollbot.telegram.callback_handler.creation import create_poll
 from pollbot.telegram.keyboard.settings import get_settings_keyboard
@@ -133,7 +134,6 @@ def handle_new_option(bot, update, session, user, text, poll, chat):
         for option in added_options:
             text += f"\n*{option}*"
         chat.send_message(text, parse_mode="markdown")
-        init_votes_for_new_options(session, poll)
     else:
         chat.send_message(i18n.t("creation.option.no_new", locale=user.locale))
 
@@ -171,15 +171,15 @@ def handle_user_option_addition(bot, update, session, user, text, poll, chat):
         user.current_poll = None
         user.expected_input = None
 
-        # Send message
+        session.commit()
+
+        # Send success message
         text = i18n.t("creation.option.multiple_added", locale=user.locale) + "\n"
         for option in added_options:
             text += f"\n*{option}*"
         chat.send_message(text, parse_mode="markdown")
 
         # Update all polls
-        init_votes_for_new_options(session, poll)
-        session.commit()
         update_poll_messages(session, bot, poll)
     else:
         chat.send_message(i18n.t("creation.option.no_new", locale=user.locale))